+++ object/activity/object_trackable.py
@@ -6,6 +6,6 @@ from object.activity.bounding_box import BoundingBox
 
 
 class ObjectTrackable:
-    def __init__(self, bounding_box: BoundingBox, data: Array[np.int]):
+    def __init__(self, bounding_box: BoundingBox, data: Array[np.uint8]):
         self.box = bounding_box
         self.data = data
