+++ object/tracking/sort_tracker.py
@@ -8,7 +8,7 @@ class SortTracker(AbstractObjectTracker):
 	def __init__(self):
 		self.__tracker = Sort()
 
-	def track(self, frame: Array[np.int], detected_boxes: List[BoundingBox]) -> List[int]:
+	def track(self, frame: Array[np.uint8], detected_boxes: List[BoundingBox]) -> List[int]:
 		rectangles = list(map(map_detected_box_to_rectangle, detected_boxes))
 		tracked_objects = self.__tracker.update(np.array(rectangles))
 		tracked_objects_ids = tracked_objects[:, 4]
